RayTracingInOneWeekend

+++ color.hpp
@@ -7,12 +7,12 @@
 
 inline void write_color(std::ostream &out, color pixel_color,
                         int samples_per_pixel) {
-    double r = pixel_color.x();
-    double g = pixel_color.y();
-    double b = pixel_color.z();
+    float r = pixel_color.x();
+    float g = pixel_color.y();
+    float b = pixel_color.z();
 
     // Divide the color by the number of samples and gamma-correct for gamma=2.0
-    const double scale = 1.0 / samples_per_pixel;
+    const float scale = 1.0 / samples_per_pixel;
     r = sqrt(scale * r);
     g = sqrt(scale * g);
     b = sqrt(scale * b);
