+++ sphere.hpp
@@ -10,19 +10,19 @@ class material;
 class sphere : public hittable {
 public:
     sphere() {}
-    sphere(point3 cen, double r, shared_ptr<material> m)
+    sphere(point3 cen, float r, shared_ptr<material> m)
         : center(cen), radius(r), mat_ptr(m){};
 
-    virtual bool hit(const ray &r, double t_min, double t_max,
+    virtual bool hit(const ray &r, float t_min, float t_max,
                      hit_record &rec) const override;
 
 public:
     point3 center;
-    double radius;
+    float radius;
     shared_ptr<material> mat_ptr;
 };
 
-inline bool sphere::hit(const ray &r, double t_min, double t_max,
+inline bool sphere::hit(const ray &r, float t_min, float t_max,
                         hit_record &rec) const {
     // t^2 b⋅b + 2tb⋅(A - C) + (A - C)⋅(A - C) - r^2 = 0
     vec3 oc = r.origin() - center;
